+++ Proyecto/Pages/AllProjects.cshtml
@@ -1,0 +1,13 @@
+﻿@page
+@model Proyecto.Pages.AllProjectsModel
+@{
+
+    ViewData ["Title"] = "All Projects";
+}
+<section id="main">
+    <h1>Details</h1>
+
+<div>
+    <h4>Todos los Proyectos a los que te puedes postular</h4>
+    
+</section>
