+++ Proyecto/Pages/Shared/_Layout.cshtml
@@ -29,13 +29,14 @@
                 <div class="navbar-collapse collapse d-sm-inline-flex flex-sm-row-reverse">
                     <ul class="navbar-nav flex-grow-1">
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-page="/Index">Home</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-page="/Privacy">Privacy</a>
+                            <a class="nav-link text-dark" asp-page="/AllProjects">All Projects</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-page="/YourProjects">Your Projects</a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-page="/Privacy">Privacy</a>
+
                         </li>
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-page="/About">About</a>
