+++ Proyecto/Pages/YourProjects.cshtml
@@ -9,6 +9,6 @@
     <h1>Details</h1>
 
 <div>
-    <h4>Project</h4>
+    <h4>Proyectos en los que estas trabajando</h4>
     
 </section>
